third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-reconnect-during.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let device, service, characteristic;
  return getHealthThermometerDeviceWithServicesDiscovered({
      filters: [{services: [health_thermometer.name]}],
  })
    .then(_ => ({device, service, characteristic} = _))
    .then(() => assert_class_string(
        characteristic, 'BluetoothRemoteGATTCharacteristic'))
    .then(() => {
      let promise = assert_promise_rejects_with_message(service.getCharacteristics(), new DOMException(
          'GATT Server is disconnected. Cannot retrieve characteristics. ' +
          '(Re)connect first with `device.gatt.connect`.',
          'NetworkError'));
      // connect() guarantees on OS-level connection, but disconnect()
      // only disconnects the current instance.
      // getHealthThermometerDeviceWithServicesDiscovered holds another
      // connection in an iframe, so disconnect() and connect() are certain to
      // reconnect.  However, disconnect() will invalidate the service object so
      // the subsequent calls made to it will fail, even after reconnecting.
      device.gatt.disconnect();
      return device.gatt.connect().then(() => promise);
    });
}, 'disconnect() and connect() called during getCharacteristics. Reject with ' +
   'NetworkError.');

</script>
